refactor: rename database follow go convention
