+++ internal/database/db.go
@@ -1,4 +1,4 @@
-package databases
+package database
 
 import (
 	"fmt"
